requirements.txt
pytrends==4.9.2
feedparser==6.0.10
google-api-python-client==2.108.0
supabase==2.3.0
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
pandas